extra pets spawned in
