+++ JS/sheep.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import { scene, earthRadius } from './globalVar.js'
+const earthRadius = 48;
 
 // 创建一个对象来表示羊
 const sheep = new THREE.Group();
@@ -54,8 +54,6 @@ sheep.position.set(2, earthRadius + 0.3, 2); // 放置在地面上
 
 // 将羊添加到场景中
 // 创建宠物网格
-const pet = sheep;
-
 sheep.scale.set(0.2, 0.2, 0.2); // 缩小羊的大小
 
 // 确保羊的所有子对象启用投射阴影
