+++ index.html
@@ -26,7 +26,7 @@
 		<button id="start-button" class="start-button">PLAY</button>
 	</div>
 	
-	<div id="game-popup" style="display: none;">
+	<div id="game-popup"  class="hidden">
 		<!-- status bar-->
 		<div id="status-bar">
 			<div>Health：<span id="life-status">100</span></div>
